+++ CHANGELOG.md
@@ -1,6 +1,9 @@
 # Changelog
 
 ## [Unreleased]
+### Removed
+- Java 8 support, by @HardNorth
+- Obsolete `UniqueID` annotation support, by @HardNorth
 
 ## [5.2.0]
 ### Fixed